requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai==0.7.0
PyPDF2==3.0.1
